% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{t_test}
\alias{t_test}
\title{R function to perform a "tidy" t-test}
\usage{
t_test(df, dv, group)
}
\arguments{
\item{df}{data frame with the dependent and group variable}

\item{dv}{raw (unquoted) name of the dependent variable}

\item{group}{raw (unquoted) name of the grouping variable}
}
\value{
A data frame containing the group means, test statistic and associated p-value, and effect size (Cohen's D)
}
\description{
R function to perform a "tidy" t-test
}
\details{
R function to compare two means using a t-test; returns the group means, test statistic and associated p-value, and effect size (Cohen's D).
}
\examples{
library(dplyr)
storms \%>\%
    filter(status \%in\% c("tropical depression", "tropical storm")) \%>\%
    mutate(category = as.integer(category)) \%>\%
    t_test(dv = category, group = status)

storms_ss <- storms \%>\%
    filter(status \%in\% c("tropical depression", "tropical storm")) \%>\%
    mutate(category = as.integer(category))

t_test_df <- t_test(storms_ss, dv = category)
t_test_df
}
